Onboarding does not discover channels but eventually node catches up
Onboarding an orderer using the last config block of the system channel.
If the system channel has channel creation transactions after the last
config block, onboarding will not discover them.

However, when the system channel is started and catches up, those
channels are discovered and pulled.

Here we have a test case to demonstrate this behavior.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I7703d28845a2a745aad7ff7cb3417d64140aaf72